+++ src/main/java/it/gov/pagopa/pu/organization/connector/pagopapayments/config/PagopaPaymentsApiClientConfig.java
@@ -1,6 +1,6 @@
 package it.gov.pagopa.pu.organization.connector.pagopapayments.config;
 
-import it.gov.pagopa.pu.organization.config.ApiClientConfig;
+import it.gov.pagopa.pu.organization.config.rest.ApiClientConfig;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import org.springframework.boot.context.properties.ConfigurationProperties;
